Testing working a lot better

+++ backend/testing/row_level/row_level_helper.js
@@ -1,58 +1,34 @@
-const { user_client } = require("../setup_fixtures");
-const testTableUpdateDelete = async ({ tableName, columnToAlter, newVal, permittedWhereClause, unpermittedWhereClause }) => {
-  const { rows: currentUserIdRows } = await user_client.query(`select current_user_id()`);
-
-  if (currentUserIdRows[0].current_user_id !== 1) {
-    debugger
-    return true
-    return 'currentUserIdRows[0].current_user_id !== 1';
-  }
-
-  //this clause should not go through (editing someone elses table)
-  badEdit = `update ${tableName} set ${columnToAlter} = ${newVal} where ${unpermittedWhereClause}`;
-  const {err} = await user_client.query(badEdit)
-  if (err){
-    return badEdit
-  }
-
-
-
-  //try delete
-  await user_client.query(`delete from ${tableName} where ${unpermittedWhereClause}`);
-
-  //check if exists
-  const { rows: invalidDeleteRows } = await user_client.query(`select * from ${tableName} where ${unpermittedWhereClause}`);
-
-  if (invalidDeleteRows[0] !== undefined) {
-    return false;
-  }
-
-  //this clause should go through (our table)
-  await user_client.query(`update ${tableName} set ${columnToAlter} = $1 where ${permittedWhereClause}`, [newVal]);
-  const { rows } = await user_client.query(`select ${columnToAlter} from ${tableName} where ${permittedWhereClause}`);
-  if (rows[0][columnToAlter] !== newVal) {
-    return false;
-  }
-
-  //try delete
-  await user_client.query(`delete from ${tableName} where ${permittedWhereClause}`);
-
-  //check if exists (shouldnt)
-  const { rows: validDeleteRows } = await user_client.query(`select * from ${tableName} where ${permittedWhereClause}`);
-  if (validDeleteRows[0] !== undefined) {
-    return false;
-  }
-  return true;
-};
-
-const testTableInsert = async ({ tableAndCols, permittedValues, unpermittedValues }) => {
+const testTableInsert = async ({ tableAndCols, permittedValues, unpermittedValues, admin_client, user_client }) => {
   await user_client.query(`insert into ${tableAndCols} values ${permittedValues}`).catch((err) => {
     expect(err).toBe(undefined);
   });
 
-  await user_client.query(`insert into ${tableAndCols} values ${unpermittedValues}`).catch((err) => {
+  const unpermittedQuery = `insert into ${tableAndCols} values ${unpermittedValues}`
+
+  //this user should not be able to insert this value
+  await user_client.query(unpermittedQuery).catch((err) => {
     expect(err).not.toBe(undefined);
   });
+
+  //insert this value in anyway with the admin client so it can be used in subsequent tests
+  await admin_client.query(`insert into ${tableAndCols} values ${unpermittedValues}`).catch((err) => {
+    expect(err).toBe(undefined);
+  });
+};
+const testTableUpdateDelete = async ({ tableName, columnToAlter, newVal, permittedWhereClause, unpermittedWhereClause, user_client }) => {
+  const rows = await user_client.query(`update ${tableName} set ${columnToAlter} = ${newVal} where ${unpermittedWhereClause} returning *`)
+  //0 rows updated (all updated rows are returned by returning *)
+  expect(rows.length).toBe(0)
+
+  const deleteRows = await user_client.query(`delete from ${tableName} where ${unpermittedWhereClause} returning *`);
+  expect(deleteRows.length).toBe(0)
+
+  const allowedRows = await user_client.query(`update ${tableName} set ${columnToAlter} = ${newVal} where ${permittedWhereClause} returning *`)
+  //altered column should be newval
+  expect(allowedRows[0][columnToAlter]).toBe(newVal)
+
+  const allowedDeleteRows = await user_client.query(`delete from ${tableName} where ${permittedWhereClause} returning *`);
+  expect(allowedDeleteRows.length).toBe(1)
 };
 
 exports.testTableUpdateDelete = testTableUpdateDelete;
